Renamed train.py to resolve conflict
ModuleNotFound error was raised from naming conflict.
